#8 fix small issues with colors, move NewProductForm's styling to NewProductForm.jsx
