+++ frontend/src/components/LandingPage.jsx
@@ -8,18 +8,6 @@ function LandingPage() {
   const [open, setOpen] = useState(false);
   const handleAddClick = () => setOpen(true);
   const handleClose = () => setOpen(false);
-  const style = {
-    position: "absolute",
-    top: "50%",
-    left: "50%",
-    transform: "translate(-50%, -50%)",
-    width: 400,
-    bgcolor: "background.paper",
-    boxShadow: 24,
-    border: "2px solid #FF7F50",
-    borderRadius: "4px",
-    p: 4,
-  };
   return (
     <Container>
       <Info />
@@ -29,9 +17,7 @@ function LandingPage() {
         </Button>
       </Box>
       <Modal open={open} onClose={handleClose}>
-        <Box sx={style} component="form">
-          <NewProductForm />
-        </Box>
+        <NewProductForm />
       </Modal>
       <AllergenProductCards />
     </Container>
